Support for creating networks with same name
Networks with the same name in different or same tenant can be created.

+++ dfa/server/dfa_fail_recovery.py
@@ -70,15 +70,17 @@ class DfaFailureRecovery(object):
         projs = self.get_fialed_projects_entries(constants.CREATE_FAIL)
         for proj in projs:
             LOG.debug("Failure recovery for project %(name)s." % (
-                          {'name': proj.name}))
+                {'name': proj.name}))
             # Try to create the project in DCNM
             try:
                 self.dcnm_client.create_project(proj.name,
-                self.cfg.dcnm.default_partition_name, proj.dci_id)
+                                                self.cfg.dcnm.
+                                                default_partition_name,
+                                                proj.dci_id)
             except dexc.DfaClientRequestFailed as e:
                 LOG.error("failure_recovery: Failed to create %(proj)s on "
                           "DCNM : %(reason)s" % (
-                          {'proj': proj.name, 'reason': str(e)}))
+                              {'proj': proj.name, 'reason': str(e)}))
             else:
                 # Request is sent successfully, update the database.
                 self.update_project_info_cache(proj.id, dci_id=proj.dci_id,
@@ -91,15 +93,17 @@ class DfaFailureRecovery(object):
         projs = self.get_fialed_projects_entries(constants.UPDATE_FAIL)
         for proj in projs:
             LOG.debug("Failure recovery for project %(name)s." % (
-                          {'name': proj.name}))
+                {'name': proj.name}))
             # This was failure of updating DCI id of the project in DCNM.
             try:
                 self.dcnm_client.update_project(proj.name,
-                self.cfg.dcnm.default_partition_name, proj.dci_id)
+                                                self.cfg.dcnm.
+                                                default_partition_name,
+                                                proj.dci_id)
             except dexc.DfaClientRequestFailed as exc:
                 LOG.error("failure_recovery: Failed to update %(proj)s on "
                           "DCNM : %(reason)s" % (
-                          {'proj': proj.name, 'reason': str(exc)}))
+                              {'proj': proj.name, 'reason': str(exc)}))
             else:
                 # Request is sent successfully, update the database.
                 self.update_project_info_cache(proj.id,
@@ -112,32 +116,33 @@ class DfaFailureRecovery(object):
         # 2. Try failure recovery for create network.
         nets = self.get_all_networks()
         for net in nets:
-            if net.result == constants.CREATE_FAIL:
+            if (net.result == constants.CREATE_FAIL
+                    and net.source.lower() == 'openstack'):
                 net_id = net.network_id
-                subnets = self.neutron_event.nclient.list_subnets(network_id=
-                                                      net_id).get('subnets')
+                subnets = self.neutron_event.nclient.list_subnets(
+                    network_id=net_id).get('subnets')
                 for subnet in subnets:
                     tenant_name = self.get_project_name(subnet['tenant_id'])
                     snet = utils.dict_to_obj(subnet)
                     try:
                         # Check if config_profile is not NULL.
                         if not net.config_profile:
-                            cfgp, fwd_mod = ( 
-                            self.dcnm_client.get_config_profile_for_network(
-                                                               net.name))
+                            cfgp, fwd_mod = (
+                                self.dcnm_client.get_config_profile_for_network(
+                                    net.name))
                             net.config_profile = cfgp
                             net.fwd_mod = fwd_mod
                         self.dcnm_client.create_network(tenant_name, net, snet)
-                    except dexc.DfaClientRequestFailed as ex:
+                    except dexc.DfaClientRequestFailed:
                         # Still is failure, only log the error.
                         emsg = 'Failed to create network %(net)s.'
                         LOG.error(emsg % {'net': net.name})
                     else:
                         # Request is sent to DCNM, update the database
-                        params = dict(columns=
-                                      dict(config_profile=net.config_profile,
-                                           fwd_mod=net.fwd_mod,
-                                           result=constants.RESULT_SUCCESS))
+                        params = dict(
+                            columns=dict(config_profile=net.config_profile,
+                                         fwd_mod=net.fwd_mod,
+                                         result=constants.RESULT_SUCCESS))
                         self.update_network(net_id, **params)
                         LOG.debug("Success on failure recovery to create "
                                   "%(net)s" % ({'net': net.name}))
@@ -161,13 +166,13 @@ class DfaFailureRecovery(object):
                 try:
                     self.neutron_event.send_vm_info(str(vm.host), str(vm_info))
                 except Exception as e:
-                    # Failed to send info to the agent. Keep the data in the 
+                    # Failed to send info to the agent. Keep the data in the
                     # database as failure to send it later.
                     LOG.error('Failed to send VM info to agent. Reason %s' % (
-                                                                str(e)))
+                        str(e)))
                 else:
-                    params = dict(columns=dict(result=
-                                               constants.RESULT_SUCCESS))
+                    params = dict(columns=dict(
+                        result=constants.RESULT_SUCCESS))
                     self.update_vm_db(vm.port_id, **params)
                     LOG.info('Created VM %(vm)s.' % {'vm': vm.name})
 
@@ -178,21 +183,22 @@ class DfaFailureRecovery(object):
                     self.neutron_event.send_vm_info(str(vm.host), str(vm_info))
                 except Exception as e:
                     LOG.error('Failed to send VM info to agent. Reason %s' % (
-                                                           str(e)))
+                        str(e)))
                 else:
                     self.delete_vm_db(vm.instance_id)
                     LOG.info('Deleted VM %(vm)s from DB.' % (
-                                           {'vm': vm.name}))
+                        {'vm': vm.name}))
 
         # 4. Try failure recovery for delete network.
         for net in nets:
-            if net.result == constants.DELETE_FAIL:
+            if (net.result == constants.DELETE_FAIL
+                    and net.source.lower() == 'openstack'):
                 net_id = net.network_id
                 segid = net.segmentation_id
                 tenant_name = self.get_project_name(net.tenant_id)
                 try:
                     self.dcnm_client.delete_network(tenant_name, net)
-                except dexc.DfaClientRequestFailed as ex:
+                except dexc.DfaClientRequestFailed:
                     # Still is failure, only log the error.
                     emsg = 'Failed to delete network %(net)s.'
                     LOG.error(emsg % {'net': net.name})
@@ -203,23 +209,24 @@ class DfaFailureRecovery(object):
                     self.delete_network_db(net_id)
                     self.segmentation_pool.add(segid)
                     LOG.debug("Success on failure recovery to deleted "
-                          "%(net)s" % ({'net': net.name}))
+                              "%(net)s" % ({'net': net.name}))
 
         # 5. Try failure recovery for delete project.
         projs = self.get_fialed_projects_entries(constants.DELETE_FAIL)
         for proj in projs:
             LOG.debug("Failure recovery for project %(name)s." % (
-                          {'name': proj.name}))
+                {'name': proj.name}))
             # Try to delete the project in DCNM
             try:
                 self.dcnm_client.delete_project(proj.name,
-                self.cfg.dcnm.default_partition_name)
+                                                self.cfg.dcnm.
+                                                default_partition_name)
             except dexc.DfaClientRequestFailed as e:
                 # Failed to delete project in DCNM.
                 # Save the info and mark it as failure and retry it later.
                 LOG.error("Failure recovery is failed to delete "
                           " %(project)s on DCNM : %(reason)s" % (
-                           {'project': proj.name, 'reason': str(e)}))
+                              {'project': proj.name, 'reason': str(e)}))
             else:
                 # Delete was successful, now update the database.
                 self.update_project_info_cache(proj.id, opcode='delete')
